added responsive styles to nav, added drawer component, extracted the links out
